backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.10.0
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
